// baml_client/types.ts
export interface SearchQueryItem {
	query_string: string;
	expected_information: string[];
}

export interface SearchResultItem {
	id: string;
	url: string;
	title?: string | null;
	source_name: string;
	snippet?: string | null;
	full_text?: string | null;
	published_date?: string | null;
	retrieval_date: string;
	author?: string | null;
	score?: number | null;
	original_query?: SearchQueryItem | null;
	metadata?: Record<string, string> | null;
}